Move NotesList tests to MealsListItem tests
